added with-base HOC
